+++ .dockerignore
@@ -1,3 +1,3 @@
 /.git
-/built
+/dist
 /node_modules
