+++ .editorconfig
@@ -1,15 +1,12 @@
 root = true
 
 [*]
-indent_style = tab
+indent_style = space
 indent_size = 2
-charset = utf-8
-insert_final_newline = true
 end_of_line = lf
+charset = utf-8
 trim_trailing_whitespace = true
+insert_final_newline = true
 
 [*.md]
 trim_trailing_whitespace = false
-
-[*.{yml,yaml}]
-indent_style = space
