+++ package.json
@@ -1,52 +1,46 @@
 {
-  "name": "@misskey-dev/summaly",
-  "version": "5.1.0",
+  "name": "@MisskeyIO/summaly",
+  "version": "5.1.1",
   "description": "Get web page's summary",
   "author": "syuilo <syuilotan@yahoo.co.jp>",
   "license": "MIT",
-  "repository": "https://github.com/misskey-dev/summaly.git",
-  "bugs": "https://github.com/misskey-dev/summaly/issues",
-  "main": "./built/index.js",
+  "repository": "https://github.com/MisskeyIO/summaly.git",
+  "bugs": "https://github.com/MisskeyIO/summaly/issues",
+  "main": "./dist/index.js",
   "type": "module",
-  "types": "./built/index.d.ts",
-  "packageManager": "pnpm@8.13.1",
-  "files": [
-    "built",
-    "LICENSE"
-  ],
+  "types": "./dist/index.d.ts",
+  "packageManager": "pnpm@9.12.3",
+  "files": ["dist", "LICENSE"],
   "scripts": {
     "build": "tsc",
-    "eslint": "eslint src --ext .js,.jsx,.ts,.tsx",
-    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --silent=false --verbose false",
-    "serve": "fastify start ./built/index.js"
+    "clean": "tsc --build --clean",
+    "format": "biome check --write",
+    "check": "tsc && biome check",
+    "test": "node --experimental-vm-modules --experimental-import-meta-resolve node_modules/jest/bin/jest.js --silent=false --verbose false",
+    "serve": "fastify start ./dist/index.js"
   },
   "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "@misskey-dev/eslint-plugin": "^1.0.0",
-    "@swc/core": "^1.3.101",
-    "@swc/jest": "^0.2.29",
-    "@types/cheerio": "0.22.18",
-    "@types/debug": "4.1.7",
-    "@types/escape-regexp": "^0.0.1",
-    "@types/node": "20.10.6",
-    "@typescript-eslint/eslint-plugin": "^6.16.0",
-    "@typescript-eslint/parser": "^6.16.0",
-    "debug": "^4.3.4",
-    "eslint": "^8.56.0",
-    "eslint-plugin-import": "^2.29.1",
-    "jest": "^29.7.0",
-    "typescript": "5.3.3"
+    "@biomejs/biome": "1.9.4",
+    "@jest/globals": "29.7.0",
+    "@swc/core": "1.9.1",
+    "@swc/jest": "0.2.37",
+    "@types/cheerio": "0.22.35",
+    "@types/debug": "4.1.12",
+    "@types/escape-regexp": "0.0.3",
+    "@types/node": "22.9.0",
+    "debug": "4.3.7",
+    "jest": "29.7.0",
+    "lefthook": "1.8.2",
+    "typescript": "5.6.3"
   },
   "dependencies": {
-    "cheerio": "1.0.0-rc.12",
+    "cheerio": "1.0.0",
     "escape-regexp": "0.0.1",
-    "fastify": "^4.25.2",
-    "fastify-cli": "^5.9.0",
-    "got": "^12.6.1",
-    "html-entities": "2.3.2",
+    "fastify": "5.0.0",
+    "fastify-cli": "7.0.1",
+    "got": "14.4.4",
+    "html-entities": "2.5.2",
     "iconv-lite": "0.6.3",
-    "jschardet": "3.0.0",
-    "private-ip": "2.3.3",
-    "trace-redirect": "1.0.6"
+    "private-ip": "3.0.2"
   }
 }
