+++ src/utils/cleanup-title.ts
@@ -1,26 +1,24 @@
 /* eslint-disable no-param-reassign */
 import escapeRegExp from 'escape-regexp';
 
-export default function(title: string, siteName?: string | null): string {
-	title = title.trim();
+export default function (_title: string, _siteName?: string | null): string {
+  const title = _title.trim();
 
-	if (siteName) {
-		siteName = siteName.trim();
+  if (_siteName) {
+    const siteName = _siteName.trim();
 
-		const x = escapeRegExp(siteName);
+    const x = escapeRegExp(siteName);
 
-		const patterns = [
-			`^(.+?)\\s?[\\-\\|:・]\\s?${x}$`,
-		];
+    const patterns = [`^(.+?)\\s?[\\-\\|:・]\\s?${x}$`];
 
-		for (let i = 0; i < patterns.length; i++) {
-			const pattern = new RegExp(patterns[i]);
-			const [, match] = pattern.exec(title) ?? [null, null];
-			if (match) {
-				return match;
-			}
-		}
-	}
+    for (const element of patterns) {
+      const pattern = new RegExp(element);
+      const [, match] = pattern.exec(title) ?? [null, null];
+      if (match) {
+        return match;
+      }
+    }
+  }
 
-	return title;
+  return title;
 }
