--- src/utils/null-or-empty.ts
@@ -1,12 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unnecessary-condition */
-export default function(val: string): boolean {
-	if (val === undefined) {
-		return true;
-	} else if (val === null) {
-		return true;
-	} else if (val.trim() === '') {
-		return true;
-	} else {
-		return false;
-	}
-}
