+++ test/tsconfig.json
@@ -1,6 +1,36 @@
 {
-	"extends": "../tsconfig.json",
-	"include": [
-		"./**/*.ts"
-	],
+  "compilerOptions": {
+    "allowJs": true,
+    "noEmitOnError": false,
+    "noImplicitAny": true,
+    "noImplicitReturns": true,
+    "noUnusedParameters": false,
+    "noUnusedLocals": false,
+    "noFallthroughCasesInSwitch": true,
+    "declaration": true,
+    "sourceMap": false,
+    "target": "es2022",
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
+    "allowSyntheticDefaultImports": true,
+    "removeComments": false,
+    "noLib": false,
+    "strict": true,
+    "strictNullChecks": true,
+    "strictPropertyInitialization": false,
+    "skipLibCheck": true,
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "baseUrl": "./",
+    "paths": {
+      "@/*": ["../src/*"]
+    },
+    "typeRoots": ["../src/@types", "../node_modules/@types", "../node_modules"],
+    "lib": ["esnext"],
+    "types": ["jest", "node"]
+  },
+  "compileOnSave": false,
+  "include": ["./**/*.ts", "../src/**/*.ts"]
 }
