+++ tsconfig.json
@@ -1,49 +1,38 @@
 {
-	"compilerOptions": {
-		"allowJs": true,
-		"noEmitOnError": false,
-		"noImplicitAny": true,
-		"noImplicitReturns": true,
-		"noUnusedParameters": false,
-		"noUnusedLocals": false,
-		"noFallthroughCasesInSwitch": true,
-		"declaration": true,
-		"sourceMap": false,
-		"target": "es2021",
-		"module": "esnext",
-		"moduleResolution": "node",
-		"allowSyntheticDefaultImports": true,
-		"removeComments": false,
-		"noLib": false,
-		"strict": true,
-		"strictNullChecks": true,
-		"strictPropertyInitialization": false,
-		"skipLibCheck": true,
-		"experimentalDecorators": true,
-		"emitDecoratorMetadata": true,
-		"resolveJsonModule": true,
-		"isolatedModules": true,
-		"rootDir": "./src",
-		"baseUrl": "./",
-		"paths": {
-			"@/*": [
-				"./src/*"
-			]
-		},
-		"outDir": "./built",
-		"types": [
-			"node"
-		],
-		"typeRoots": [
-			"./node_modules/@types",
-			"./src/@types"
-		],
-		"lib": [
-			"esnext"
-		]
-	},
-	"compileOnSave": false,
-	"include": [
-		"./src/**/*.ts"
-	],
+  "compilerOptions": {
+    "allowJs": true,
+    "noEmitOnError": false,
+    "noImplicitAny": true,
+    "noImplicitReturns": true,
+    "noUnusedParameters": false,
+    "noUnusedLocals": false,
+    "noFallthroughCasesInSwitch": true,
+    "declaration": true,
+    "sourceMap": false,
+    "target": "es2022",
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
+    "allowSyntheticDefaultImports": true,
+    "removeComments": false,
+    "noLib": false,
+    "strict": true,
+    "strictNullChecks": true,
+    "strictPropertyInitialization": false,
+    "skipLibCheck": true,
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "rootDir": "./src",
+    "baseUrl": "./",
+    "paths": {
+      "@/*": ["./src/*"]
+    },
+    "outDir": "./dist",
+    "types": ["node"],
+    "typeRoots": ["./src/@types", "./node_modules/@types", "./node_modules"],
+    "lib": ["esnext"]
+  },
+  "compileOnSave": false,
+  "include": ["./src/**/*.ts"]
 }
